requirements.txt
aldryn-boilerplates==0.7.7
-e git://github.com/aldryn/aldryn-forms.git@2932c85ffc87717ed1b9a3cdf1ada0e3ea2cdbb0#egg=aldryn_forms
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
cssselect==1.0.3
cssutils==1.0.2
decorator==4.3.0
dj-database-url==0.5.0
Django==1.11.13
django-absolute==0.3
django-admin-sortable2==0.6.19
django-appconf==1.0.2
django-classy-tags==0.8.0
-e git://github.com/divio/django-cms.git@2be4e359bfc9049ab82479add995a05c324b41fa#egg=django_cms
django-emailit==0.2.3
django-filer==1.3.2
django-formtools==2.1
django-fsm==2.6.0
django-mptt==0.8.7
django-polymorphic==2.0.2
django-ranged-response==0.2.0
django-sekizai==0.10.0
django-simple-captcha==0.5.9
django-sizefield==0.9.1
django-su==0.6.0
django-tablib==3.2
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-helper==1.1.0
djangocms-installer==1.0.1
djangocms-link==2.1.2
-e git+https://github.com/divio/djangocms-moderation.git@4e9f0d543c9fd2587e62fe81ceab212e32fd3cb0#egg=djangocms_moderation
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
-e git://github.com/divio/djangocms-versioning.git@f05eed311cf0bca4a3fc1d6005d682ecc847ec80#egg=djangocms_versioning
djangocms-video==2.0.4
docopt==0.6.2
easy-thumbnails==2.5
et-xmlfile==1.0.1
flake8==3.5.0
Faker==0.9.0 
factory-boy==2.11.1 
python-dateutil==2.7.3 
text-unidecode==1.2
html5lib==0.9999999
idna==2.7
ipdb==0.11
iptools==0.6.1
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jdcal==1.4
jedi==0.12.0
lxml==4.2.1
mccabe==0.6.1
mock==2.0.0
odfpy==1.3.6
openpyxl==2.4.9
parso==0.2.1
pbr==4.1.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.6.0
premailer==3.2.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
pycodestyle==2.3.1
pyflakes==2.0.0
Pygments==2.2.0
pytz==2018.4
PyYAML==3.12
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
tablib==0.12.1
tox==3.0.0
traitlets==4.3.2
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
